controller of brandamount and inventory

+++ Data/Entities/BrandInventory.cs
@@ -7,12 +7,12 @@ public class BrandInventory
    
         public long Id { get; set; }
         public int Count { get; set; }
-        public long BrandId { get; set; }
+        public int BrandId { get; set; }
         
         [JsonIgnore]
         public Brand Brand { get; set; }
       
-        public long DoctorId { get; set; }
+        public int DoctorId { get; set; }
          [JsonIgnore]
         public Doctor Doctor { get; set; }
       //  public virtual ICollection<Vaccine> VaccineName { get; set; }
